apps/inspiration-organizer/requirements.txt
numpy==1.24.3
pandas==2.0.2
nltk==3.8.1
whoosh==2.7.4
Pillow==10.0.0
librosa==0.10.0.post2
opencv-python==4.8.0.74
python-dotenv==1.0.0
openai==0.27.8
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.1
seaborn==0.12.2
